simple-types: fix type propagation
Provides fix to module simple-types that analysed and propagated
type of each object throught the LLVM IR module. This propagation
was done by analyzing priorities of each object and making
debug information having the heighest priority. This behaviour
is not always inteded as for example this would propagate
base type of object to every usage of it which would result in
loss of cast to the higher type that is oftend needed in
various situastions.

This commit alterns algorithm that analyse type that will be propagate throught
usage to be the heighest priority type -> with int the largest int type and so on.
Priority of type si obtained from method getHigherPriorityType.