capstone2llvmir: arm64: store to FP register requires different ct
Currently ther is not a good design of storing operands into
other operands (like it is with loading operands). As there may
occour errors f.e. on side of capstone we have to make sure that
operands are converted correctly. This is only temprorary solution
as better refactorization of store operands method is required.
This hovewer affects all architectures and design shuld be better
rethought.